requirements.txt
# LLM and AI - Using latest modular LangChain packages with free Hugging Face embeddings
langchain-core>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.38
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Web Scraping
ddgs>=5.0.0  # Renamed from duckduckgo-search
beautifulsoup4>=4.12.0
requests>=2.31.0

# Excel and Data Processing
openpyxl>=3.1.2  # Required for pandas Excel read/write
pandas>=2.2.0

# Microsoft Graph API
msal>=1.25.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Logging and Monitoring
coloredlogs>=15.0

# Web Interface
streamlit>=1.30.0
